Refactoring
SE_ManagerのPlay関数の引数の型を変更しました。

+++ JumperWireGame/Assets/Scripts/Main_Scene/Ball_Generator.cs
@@ -125,7 +125,7 @@ public class Ball_Generator : MonoBehaviour
                 break;
         }
 
-        sePlayer.Play((int)SE_Manager.kind.Generate_balls);
+        sePlayer.Play(SE_Manager.kind.Generate_balls);
         scoreManager.AddGeneratedBall();
     }
 }
