+++ JumperWireGame/Assets/Scripts/Main_Scene/Capsule_Manager.cs
@@ -20,6 +20,6 @@ public class Capsule_Manager : Ball_Manager
     }
 
     protected override void PlayOKSound(){
-        sePlayer.Play((int)SE_Manager.kind.OK_capsule);
+        sePlayer.Play(SE_Manager.kind.OK_capsule);
     }
 }
